Extract comms-related functionality into `AbstractSendable`
This creates a place for communication methods to be centralized and to
provide default definitions for them that can be refined as necessary
in sub-classes.
