+++ syft/generic/abstract/tensor.py
@@ -3,12 +3,12 @@ import weakref
 
 import syft as sy
 from syft.generic.abstract.object import _apply_args  # noqa: F401
-from syft.generic.abstract.object import AbstractObject
+from syft.generic.abstract.sendable import AbstractSendable
 from syft.generic.abstract.object import initialize_object
 from syft.serde.syft_serializable import SyftSerializable
 
 
-class AbstractTensor(AbstractObject, SyftSerializable):
+class AbstractTensor(AbstractSendable, SyftSerializable):
     def __init__(
         self,
         id: int = None,
