+++ syft/generic/pointers/object_pointer.py
@@ -11,7 +11,7 @@ from syft.generic.frameworks.hook.hook_args import register_forward_func
 from syft.generic.frameworks.hook.hook_args import register_backward_func
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.types import FrameworkTensor
-from syft.generic.abstract.object import AbstractObject
+from syft.generic.abstract.sendable import AbstractSendable
 from syft.messaging.message import ForceObjectDeleteMessage
 from syft.workers.abstract import AbstractWorker
 
@@ -24,7 +24,7 @@ if TYPE_CHECKING:
     from syft.workers.base import BaseWorker
 
 
-class ObjectPointer(AbstractObject, SyftSerializable):
+class ObjectPointer(AbstractSendable, SyftSerializable):
     """A pointer to a remote object.
 
     An ObjectPointer forwards all API calls to the remote. ObjectPointer objects
