+++ syft/generic/string.py
@@ -4,13 +4,13 @@ from typing import Union
 import syft as sy
 from syft.generic.pointers.string_pointer import StringPointer
 from syft.workers.base import BaseWorker
-from syft.generic.abstract.object import AbstractObject
+from syft.generic.abstract.sendable import AbstractSendable
 from syft.generic.frameworks.overload import overloaded
 from syft.generic.frameworks.hook import hook_args
 from syft_proto.generic.string_pb2 import String as StringPB
 
 
-class String(AbstractObject):
+class String(AbstractSendable):
     """
        This is a class that wraps the Python built-in `str` class. In addition to
     providing access to most of `str`'s method call API, it allows sending
